+++ integration/playwright-web/src/Actors.ts
@@ -1,5 +1,6 @@
 import { Actor, Cast, TakeNotes } from '@serenity-js/core';
 import { BrowseTheWebWithPlaywright, PlaywrightOptions } from '@serenity-js/playwright';
+import { CallAnApi } from '@serenity-js/rest';
 import * as playwright from 'playwright-core';
 
 export class Actors implements Cast {
@@ -10,13 +11,10 @@ export class Actors implements Cast {
     }
 
     prepare(actor: Actor): Actor {
-        if (actor.name === `Adam who can't browse the web`) {
-            return actor;
-        }
-
         return actor.whoCan(
             BrowseTheWebWithPlaywright.using(this.browser, this.options),
             TakeNotes.usingAnEmptyNotepad(),
+            CallAnApi.at(this.options.baseURL)
         );
     }
 }