+++ integration/web-specs/spec/stage/crew/photographer/strategies/TakePhotosOfInteractions.spec.ts
@@ -2,7 +2,12 @@ import 'mocha';
 
 import { EventRecorder, expect, PickEvent } from '@integration/testing-tools';
 import { Duration } from '@serenity-js/core';
-import { ActivityRelatedArtifactGenerated, ActivityStarts, InteractionFinished, InteractionStarts, SceneFinishes, SceneStarts } from '@serenity-js/core/lib/events';
+import {
+    ActivityRelatedArtifactGenerated,
+    ActivityStarts,
+    SceneFinishes,
+    SceneStarts
+} from '@serenity-js/core/lib/events';
 import { CorrelationId, Photo } from '@serenity-js/core/lib/model';
 import { Stage } from '@serenity-js/core/lib/stage';
 import { BrowseTheWeb, Photographer, TakePhotosOfInteractions } from '@serenity-js/web';
@@ -126,18 +131,5 @@ describe('Photographer', () => {
                     );
                 });
         });
-
-        it(`does not attempt to take a photo if the actor in the spotlight can't BrowseTheWeb`, () =>
-            expect(stage.theActorCalled(`Adam who can't browse the web`).attemptsTo(
-                Perform.interactionThatSucceeds(1),
-            )).to.be.fulfilled.then(() => stage.waitForNextCue().then(() => {
-
-                expect(recorder.events.length).to.equal(3);
-                expect(recorder.events[0]).to.be.instanceOf(SceneStarts);
-                expect(recorder.events[1]).to.be.instanceOf(InteractionStarts);
-                expect(recorder.events[2]).to.be.instanceOf(InteractionFinished);
-
-                // no artifacts generated for an actor with no ability to BrowseTheWeb
-            })));
     });
 });