+++ packages/serenity-bdd/src/cli/stage/UpdateCommandActors.ts
@@ -1,8 +1,8 @@
-import type { Actor, Cast} from '@serenity-js/core';
+import type { Actor, Cast } from '@serenity-js/core';
 import { TakeNotes } from '@serenity-js/core';
 import type { Path } from '@serenity-js/core/lib/io';
+import type { AxiosRequestConfigDefaults} from '@serenity-js/rest';
 import { CallAnApi } from '@serenity-js/rest';
-import type { AxiosInstance } from 'axios';
 
 import { UseFileSystem } from '../screenplay';
 
@@ -12,13 +12,13 @@ import { UseFileSystem } from '../screenplay';
 export class UpdateCommandActors implements Cast {
     constructor(
         private readonly cwd: Path,
-        private readonly createAxios: () => AxiosInstance,
+        private readonly axiosConfigDefaults: AxiosRequestConfigDefaults,
     ){
     }
 
     prepare(actor: Actor): Actor {
         return actor.whoCan(
-            CallAnApi.using(this.createAxios()),
+            CallAnApi.using(this.axiosConfigDefaults),
             UseFileSystem.at(this.cwd),
             TakeNotes.usingAnEmptyNotepad(),
         );